+++ service/model_meta.go
@@ -30,7 +30,7 @@ type WaitSyncMeta struct {
 	SourceSchemaName string `gorm:"not null;index:idx_schema_table_mode,unique;comment:'源端 schema'" json:"source_schema_name"`
 	SourceTableName  string `gorm:"not null;index:idx_schema_table_mode,unique;comment:'源端表名'" json:"source_table_name"`
 	SyncMode         string `gorm:"not null;index:idx_schema_table_mode,unique;comment:'同步模式'" json:"sync_mode"`
-	FullGlobalSCN    int    `gorm:"comment:'全量全局 SCN'" json:"full_global_scn"`
+	FullGlobalSCN    uint64 `gorm:"comment:'全量全局 SCN'" json:"full_global_scn"`
 	FullSplitTimes   int    `gorm:"comment:'全量任务切分 SQL 次数'" json:"full_split_times"`
 	IsPartition      string `gorm:"comment:'是否是分区表'" json:"is_partition"` // 同步转换统一转换成非分区表，此处只做标志
 	BaseModel
@@ -41,7 +41,7 @@ type FullSyncMeta struct {
 	ID               uint   `gorm:"primary_key;autoIncrement;comment:'自增编号'" json:"id"`
 	SourceSchemaName string `gorm:"not null;index:idx_schema_table_rowid;comment:'源端 schema'" json:"source_schema_name"`
 	SourceTableName  string `gorm:"not null;index:idx_schema_table_rowid;comment:'源端表名'" json:"source_table_name"`
-	GlobalSCN        int    `gorm:"comment:'全局 SCN'" json:"global_scn"`
+	GlobalSCN        uint64 `gorm:"comment:'全局 SCN'" json:"global_scn"`
 	SourceSQL        string `gorm:"type:text;comment:'源端 sql'" json:"source_sql"`
 	RowidSQL         string `gorm:"type:varchar(300);not null;index:idx_schema_table_rowid;comment:'表 rowid 切分SQL'" json:"rowid_sql"`
 	IsPartition      string `gorm:"comment:'是否是分区表'" json:"is_partition"` // 同步转换统一转换成非分区表，此处只做标志
@@ -52,10 +52,10 @@ type FullSyncMeta struct {
 // 增量同步元数据表
 type IncrementSyncMeta struct {
 	ID               uint   `gorm:"primary_key;autoIncrement;comment:'自增编号'" json:"id"`
-	GlobalSCN        int    `gorm:"comment:'全局 SCN'" json:"global_scn"`
+	GlobalSCN        uint64 `gorm:"comment:'全局 SCN'" json:"global_scn"`
 	SourceSchemaName string `gorm:"not null;index:unique_schema_table,unique;comment:'源端 schema'" json:"source_schema_name"`
 	SourceTableName  string `gorm:"not null;index:unique_schema_table,unique;comment:'源端表名'" json:"source_table_name"`
-	SourceTableSCN   int    `gorm:"comment:'表同步 SCN'" json:"source_table_scn"`
+	SourceTableSCN   uint64 `gorm:"comment:'表同步 SCN'" json:"source_table_scn"`
 	IsPartition      string `gorm:"comment:'是否是分区表'" json:"is_partition"` // 同步转换统一转换成非分区表，此处只做标志
 	BaseModel
 }
@@ -82,7 +82,7 @@ func (f *IncrementSyncMeta) GetIncrementSyncMetaRecordCounts(schemaName, tableNa
 	return int(count), nil
 }
 
-func (e *Engine) UpdateTableIncrementMetaALLSCNRecord(sourceSchemaName, sourceTableName, operationType string, globalSCN, sourceTableSCN int) error {
+func (e *Engine) UpdateTableIncrementMetaALLSCNRecord(sourceSchemaName, sourceTableName, operationType string, globalSCN, sourceTableSCN uint64) error {
 	if operationType == utils.DropTableOperation {
 		if err := e.GormDB.Transaction(func(tx *gorm.DB) error {
 			if err := tx.Where("source_schema_name = ? and source_table_name = ?",
@@ -114,8 +114,8 @@ func (e *Engine) UpdateTableIncrementMetaALLSCNRecord(sourceSchemaName, sourceTa
 }
 
 func (e *Engine) UpdateSingleTableIncrementMetaSCNByCurrentRedo(
-	sourceSchemaName string, lastRedoLogMaxSCN, logFileStartSCN, logFileEndSCN int) error {
-	var logFileSCN int
+	sourceSchemaName string, lastRedoLogMaxSCN, logFileStartSCN, logFileEndSCN uint64) error {
+	var logFileSCN uint64
 	if logFileEndSCN >= lastRedoLogMaxSCN {
 		logFileSCN = logFileStartSCN
 	} else {
@@ -146,8 +146,8 @@ func (e *Engine) UpdateSingleTableIncrementMetaSCNByCurrentRedo(
 }
 
 func (e *Engine) UpdateSingleTableIncrementMetaSCNByNonCurrentRedo(
-	sourceSchemaName string, lastRedoLogMaxSCN, logFileStartSCN, logFileEndSCN int, transferTableSlice []string) error {
-	var logFileSCN int
+	sourceSchemaName string, lastRedoLogMaxSCN, logFileStartSCN, logFileEndSCN uint64, transferTableSlice []string) error {
+	var logFileSCN uint64
 	if logFileEndSCN >= lastRedoLogMaxSCN {
 		logFileSCN = logFileStartSCN
 	} else {
@@ -169,7 +169,7 @@ func (e *Engine) UpdateSingleTableIncrementMetaSCNByNonCurrentRedo(
 }
 
 func (e *Engine) UpdateSingleTableIncrementMetaSCNByArchivedLog(
-	sourceSchemaName string, logFileEndSCN int, transferTableSlice []string) error {
+	sourceSchemaName string, logFileEndSCN uint64, transferTableSlice []string) error {
 	for _, table := range transferTableSlice {
 		if err := e.GormDB.Model(IncrementSyncMeta{}).Where(
 			"source_schema_name = ? and source_table_name = ?",
